+++ spec/kubernetes_template_rendering/template_directory_renderer_spec.rb
@@ -21,13 +21,19 @@ RSpec.describe KubernetesTemplateRendering::TemplateDirectoryRenderer do
     FileUtils.rm_r(template_directory)
   end
 
-  { "test": "test", "prod.gcp": "prod", "SPP-PLACEHOLDER": "staging", "SPP-PLACEHOLDER.eu": "staging" }.each do |resource_definition_name, expected_kubernetes_cluster_type|
+  {
+    "test"               => { cluster_type: "test",    spp: false },
+    "prod.gcp"           => { cluster_type: "prod",    spp: false },
+    "staging"            => { cluster_type: "staging", spp: false },
+    "SPP-PLACEHOLDER"    => { cluster_type: "staging", spp: true },
+    "SPP-PLACEHOLDER.eu" => { cluster_type: "staging", spp: true }
+  }.each do |resource_definition_name, expected|
     it "builds a ResourceSet with an appropriate kubernetes_cluster_type for each directory and calls render on it" do
-      definition = build_definition(name: resource_definition_name.to_s, rendered_directory: rendered_directory, template_directory: template_directory, variables: variables)
+      definition = build_definition(name: resource_definition_name, rendered_directory: rendered_directory, template_directory: template_directory, variables: variables)
       definitions_path = File.join(template_directory, described_class::DEFINITIONS_FILENAME)
       File.write(definitions_path, definition.to_yaml)
 
-      expected_config = OpenStruct.new(definition[resource_definition_name.to_s])
+      expected_config = OpenStruct.new(definition[resource_definition_name])
       resource_set = instance_double(KubernetesTemplateRendering::ResourceSet)
 
       expect(resource_set).to receive(:render)
@@ -38,7 +44,8 @@ RSpec.describe KubernetesTemplateRendering::TemplateDirectoryRenderer do
             rendered_directory: rendered_directory,
             template_directory: template_directory,
             definitions_path: definitions_path,
-            kubernetes_cluster_type: expected_kubernetes_cluster_type,
+            kubernetes_cluster_type: expected[:cluster_type],
+            spp: expected[:spp],
             variable_overrides: {},
             source_repo: nil
           )